fix: Use 1 conf file to avoid importing every fixture
