+++ tests/conftest.py
@@ -7,8 +7,17 @@ from httpx import AsyncClient
 from main import app, config, db
 from main.libs.log import get_logger
 from main.models.base import BaseModel
+from main.models.category import CategoryModel
+from main.models.user import UserModel
+from main.utils.auth import create_access_token_from_id
 
-from .fixtures import access_token, category, item, user
+from .helpers import (
+    mock_email,
+    mock_password,
+    prepare_category,
+    prepare_item,
+    prepare_user,
+)
 
 logger = get_logger(__name__)
 
@@ -54,13 +63,21 @@ async def client():
         yield client
 
 
-__all__ = [
-    "event_loop",
-    "recreate_database",
-    "database",
-    "client",
-    "user",
-    "category",
-    "item",
-    "access_token",
-]
+@pytest.fixture
+async def user():
+    return await prepare_user(email=mock_email, password=mock_password)
+
+
+@pytest.fixture
+async def category(user: UserModel):
+    return await prepare_category(user.id)
+
+
+@pytest.fixture
+async def item(user: UserModel, category: CategoryModel):
+    return await prepare_item(category_id=category.id, creator_id=user.id)
+
+
+@pytest.fixture
+def access_token(user: UserModel):
+    return create_access_token_from_id(user.id)
